adds abstract observer, refactors integrity updates to observer pattern and updates TimerTestScene to match new refactored code
