+++ Assets/MyPrefabs/Scripts/Game/Observer.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 06ceb73ffb46470182791aa19dd109f1
+timeCreated: 1700592664
