+++ Assets/MyPrefabs/Scripts/Game/Observer/AbstractObserver.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 3b280069a64e4a7c8ff4da7ba5141c03
+timeCreated: 1700592702
