+++ Assets/MyPrefabs/Scripts/Game/Observer/GameTaskObserver.cs.meta
