+++ Assets/MyPrefabs/Scripts/Game/Observer/IntegrityObserver.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: a0c1c7f01acf4a07b37368f229a4090a
+timeCreated: 1700591761
