+++ Assets/MyPrefabs/Scripts/Game/Tasks/GameTask.cs
@@ -18,8 +18,7 @@ namespace MyPrefabs.Scripts.Game.Tasks
         [Delayed]
         public string taskName;
 
-        [SerializeField] private int integrityValue = 5;
-        private Integrity _integrity;
+        [SerializeField] public int integrityValue = 5;
 
         protected TaskState currentTaskState;
         public event Action<GameTask> TaskSuccessful;
@@ -54,11 +53,6 @@ namespace MyPrefabs.Scripts.Game.Tasks
         /// </summary>
         protected abstract void AfterStateCheck();
 
-        private void Start()
-        {
-            _integrity = FindObjectOfType<Integrity>();
-        }
-
         /// <summary>
         /// Template method based update procedure for processing game task logic and updating game task state
         /// </summary>
@@ -80,11 +74,9 @@ namespace MyPrefabs.Scripts.Game.Tasks
             {
                 case TaskState.Failed:
                     TaskFailed?.Invoke(this);
-                    _integrity.DecrementIntegrity(integrityValue);
                     break;
                 case TaskState.Successful:
                     TaskSuccessful?.Invoke(this);
-                    _integrity.IncrementIntegrity(integrityValue);
                     break;
                 case TaskState.Ongoing:
                     break;
